pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "Telegram bot for TP Link router"
authors = ["Emil Abdulatypov <emiaesthetic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
isort = "^5.12.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pytest = "^7.4.0"
aiogram = "3.0.0rc2"


[tool.black]
line-length = 80
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
